code duplication was removed
now UnitFragment from free and paid version extend abstract class UnitFragmentBase from main
