+++ Lab_1/app/src/free/java/by/bsuir/wildboom/lab1/UnitFragment.java
@@ -1,119 +1,16 @@
 package by.bsuir.wildboom.lab1;
 
-import android.R.layout;
 import android.os.Bundle;
-import android.text.InputType;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.EditText;
-import android.widget.Spinner;
 
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
+import androidx.annotation.NonNull;
 
-public class UnitFragment extends Fragment {
-    EditText lineTop;
-    EditText lineBottom;
-    Spinner spinnerTop;
-    Spinner spinnerBottom;
-
-    ConversionViewModel conversionViewModel;
-    SharedViewModel sharedViewModel;
-
-    public UnitFragment() {
-        // Required empty public constructor
-    }
-
+public class UnitFragment extends UnitFragmentBase {
     @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_unit, container, false);
-
-        lineTop = view.findViewById(R.id.textView_top);
-        lineTop.setInputType(InputType.TYPE_NULL);
-        lineBottom = view.findViewById(R.id.textView_bottom);
-        lineBottom.setInputType(InputType.TYPE_NULL);
-        spinnerTop = view.findViewById(R.id.spinner_top);
-        spinnerBottom = view.findViewById(R.id.spinner_bottom);
-
-        conversionViewModel = new ViewModelProvider(requireActivity()).get(ConversionViewModel.class);
-        sharedViewModel = new ViewModelProvider(requireActivity()).get(SharedViewModel.class);
-        sharedViewModel.select("");
-
-        conversionViewModel.getSelectedValue().observe(getViewLifecycleOwner(), value -> lineTop.setText(value));
-        sharedViewModel.getSelected().observe(getViewLifecycleOwner(), this::setNewVal);
-
-        conversionViewModel.getSelectedSpinnerTop().observe(getViewLifecycleOwner(),
-                value -> {
-                    String[] values = conversionViewModel.conversion.getValues();
-                    ArrayAdapter<String> adapterTop = new ArrayAdapter<>(
-                            requireActivity().getBaseContext(),
-                            layout.simple_spinner_item, values);
-                    adapterTop.setDropDownViewResource(layout.simple_spinner_dropdown_item);
-                    spinnerTop.setAdapter(adapterTop);
-                    spinnerTop.setSelection(value);
-                });
-
-        conversionViewModel.getSelectedSpinnerBottom().observe(getViewLifecycleOwner(),
-                value -> {
-                    String[] values = conversionViewModel.conversion.getValues();
-                    ArrayAdapter<String> adapterBottom = new ArrayAdapter<>(
-                            requireActivity().getBaseContext(),
-                            layout.simple_spinner_item, values);
-                    adapterBottom.setDropDownViewResource(layout.simple_spinner_dropdown_item);
-                    spinnerBottom.setAdapter(adapterBottom);
-                    spinnerBottom.setSelection(value);
-                });
-
-        spinnerTop.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener()
-        {
-            @Override
-            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
-                conversionViewModel.selectSpinnerTop(position);
-                lineBottom.setText(conversionViewModel.makeConversion(lineTop.getText().toString()));
-            }
-            @Override
-            public void onNothingSelected(AdapterView<?> arg0) {
-            }
-        });
-
-        spinnerBottom.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener()
-        {
-            @Override
-            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
-                conversionViewModel.selectSpinnerBottom(position);
-                lineBottom.setText(conversionViewModel.makeConversion(lineTop.getText().toString()));
-            }
-            @Override
-            public void onNothingSelected(AdapterView<?> arg0) {
-            }
-        });
-
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View view = onCreateViewHelp(inflater, container);
         return view;
-    }
-
-    public void setNewVal(String item) {
-        String lineValue = lineTop.getText().toString();
-        if (item != null) {
-            if (!item.equals("|"))
-                lineValue += item;
-            else {
-                String answer = "";
-                if (lineValue.length() != 0) {
-                    answer = lineValue.substring(0, lineValue.length() - 1);
-                }
-                lineValue = answer;
-            }
-        }
-        lineTop.setText(lineValue);
-        String text_down = conversionViewModel.selectValue(lineValue);
-        lineBottom.setText(text_down);
     }
 }
